Added token expiry verification
